+++ config/auth_u2m_test.go
@@ -8,9 +8,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/databricks/databricks-sdk-go/credentials/cache"
+	"github.com/databricks/databricks-sdk-go/credentials/oauth"
 	"github.com/databricks/databricks-sdk-go/httpclient/fixtures"
-	"github.com/databricks/databricks-sdk-go/internal/credentials/cache"
-	"github.com/databricks/databricks-sdk-go/internal/credentials/oauth"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/oauth2"
 )
